// src/models/auth.model.ts
export interface User {
    id: number;
    dateCreated: number;
    username: string;
    password: string;
}

export interface Session {
    id: number;
    dateCreated: number;
    username: string;
    /**
     * Timestamp indicating when the session was created, in Unix milliseconds.
     */
    issued: number;
    /**
     * Timestamp indicating when the session should expire, in Unix milliseconds.
     */
    expires: number;
}

/**
 * Identical to the Session type, but without the `issued` and `expires` properties.
 */
export type PartialSession = Omit<Session, "issued" | "expires">;

export interface EncodeResult {
    token: string;
    refreshToken: string;
    expires: number;
    issued: number;
    refreshTokenExpiresAt: number;
}

export type DecodeResult =
    | {
    type: "valid";
    session: Session;
}
    | {
    type: "integrity-error";
}
    | {
    type: "invalid-token";
};

export type ExpirationStatus = "expired" | "active" | "grace";


export interface IJWTSession extends Session {

}


export interface IAuthLoginRequest {
    email: string;
    password: string;
    rememberMe: boolean;
}